requirements.txt
Flask
pandas
python-dotenv
google-generativeai
gunicorn
SQLAlchemy
psycopg2-binary